gather row types
